feat/performance
- Defaults: output file type -> PNG
- Performance: Processing speed increased by 50% for 4k images
  Benchmark: Old: 2.0s, New: 1.33s (Proc time for -t all)
Achieved by pre-casting palettes to f32
- UX/Output: Improved -t output with better formatting and included performance measurements

+++ src/config.rs
@@ -10,7 +10,7 @@ pub struct Config {
     pub output: String,
     //pub format: String,
 
-    pub palette: Vec<[u8; 3]>,
+    pub palette: Vec<[f32; 3]>,
     pub palette_name: String,
     pub dither: Dither,
     pub ampl: f32,
